docs and errors
