+++ bindings/java/src/main/java/org/apache/opendal/AsyncExecutor.java
@@ -19,6 +19,11 @@
 
 package org.apache.opendal;
 
+/**
+ * AsyncExecutor represents an underneath OpenDAL executor that runs async tasks spawned in the Rust world.
+ *
+ * <p>If the executor is passed to construct operators, the executor must outlive the operators.</p>
+ */
 public class AsyncExecutor extends NativeObject {
     public static AsyncExecutor createTokioExecutor(int cores) {
         return new AsyncExecutor(makeTokioExecutor(cores));
